refactor(board): move hand zone to hand directory
