+++ src/app/app.component.html
@@ -1,3 +1,3 @@
-<div class="min-h-screen place-content-center bg-gray-800 text-gray-200">
+<div class="min-h-screen bg-gray-800 text-gray-200">
   <router-outlet />
 </div>
