--- src/app/components/game-zones/hand-zone/index.ts
@@ -1,2 +1,0 @@
-export * from './hand-zone.component';
-export { DisplayHandService } from './services';
