+++ src/app/components/game-zones/hand/hand-zone/hand-zone.component.html
