+++ src/app/components/game-zones/hand/hand-zone/index.ts
@@ -1,0 +1,1 @@
+export * from './hand-zone.component';
